requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# Image processing
opencv-python==4.9.0.80
Pillow==10.4.0
numpy==1.26.4
scikit-image>=0.23.0

# OCR
paddlepaddle==3.0.0
paddleocr==2.9.1

# ML / embeddings
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
open-clip-torch==2.24.0

# Barcode
pyzbar==0.1.9
python-barcode==0.15.1

# DB / storage
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Async task queue
celery==5.4.0
redis==5.0.4

# Utils
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose==3.3.0
httpx==0.27.0
loguru==0.7.2
asyncpg==0.29.0